feat: discord error bot is back
